+++ RELEASES.md
@@ -72,6 +72,7 @@ Upcoming
 - Add defaultValueMethods to Parameter
 - Don't use ConversationId from previous sender response
 - Add GetPrincipalPipe and IsUserInRolePipe to stub4testtool.xsl
+- Add MessageStoreSender and MessageStoreListener